+++ tests/Language/GraphQL/AST/ParserSpec.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
-module Language.GraphQL.ParserSpec
+module Language.GraphQL.AST.ParserSpec
     ( spec
     ) where
 
-import Language.GraphQL.Parser (document)
+import Language.GraphQL.AST.Parser
 import Test.Hspec (Spec, describe, it)
 import Test.Hspec.Megaparsec (shouldSucceedOn)
 import Text.Megaparsec (parse)